+++ simple_model/models.py
@@ -1,15 +1,13 @@
 from typing import Any, Callable, Iterable, Iterator, Tuple, Union
 
-from .base import BaseModel
+from .base import ModelMetaClass
 from .exceptions import ValidationError
 from .fields import ModelField
 from .utils import getkey
 
 
-class Model(metaclass=BaseModel):
+class BaseModel:
     def __init__(self, **kwargs):
-        self._is_valid = False
-
         for field_name in self._meta.fields:
             descriptor = self._meta.descriptors[field_name]
 
@@ -18,7 +16,10 @@ class Model(metaclass=BaseModel):
             factory = default if isinstance(default, Callable) else None
             field_value = factory() if factory and not field_value else kwargs.get(field_name, default)
 
-            setattr(self, field_name, field_value)
+            if descriptor and descriptor.is_property:
+                self.__setattr__(field_name, field_value)
+            else:
+                object.__setattr__(self, field_name, field_value)
 
         self.__post_init__(**kwargs)
 
@@ -62,9 +63,9 @@ class Model(metaclass=BaseModel):
         try:
             super().__setattr__(name, value)
         except AttributeError:
-            descriptor = self._meta.descriptors.get(name)
+            descriptor = self._meta.descriptors.get(name, False)
             if descriptor and not descriptor.is_property:
-                raise
+                raise  # TODO: find/implement a test case for this statement
 
     def _get_fields(self) -> Iterator[Tuple[str, ModelField]]:
         return (
@@ -95,19 +96,19 @@ class Model(metaclass=BaseModel):
             if descriptor.is_property:
                 continue
 
-            value = getattr(self, name)
+            value = object.__getattribute__(self, name)
             new_value = descriptor.convert_to_type(self, value)
-            setattr(self, name, new_value)
+
+            object.__setattr__(self, name, new_value)
 
     def validate(self, raise_exception: bool = True) -> Union[None, bool]:
-        self._is_valid = True
         self.convert_fields()
 
         for name, descriptor in self._get_fields():
             if descriptor.is_property and descriptor._validate is None:
                 continue
 
-            value = getattr(self, name)
+            value = object.__getattribute__(self, name)
             try:
                 value = descriptor.validate(self, value)
             except ValidationError:
@@ -119,13 +120,61 @@ class Model(metaclass=BaseModel):
                 self._is_valid = False
                 raise
 
-            setattr(self, name, value)
+            try:
+                object.__setattr__(self, name, value)
+            except AttributeError:
+                self.__setattr__(name, value)
 
+        self._is_valid = True
         return None if raise_exception else True
 
     def as_dict(self):
         """
         Returns the model as a dict
         """
+        from .converters import to_dict
+        return to_dict(self)
+
+
+class Model(BaseModel, metaclass=ModelMetaClass):
+    pass
+
+
+class LazyModel(BaseModel, metaclass=ModelMetaClass):
+    """ LazyModel is a short name for LazyValidationModel. Example:
+
+    class FooModel(LazyModel):
+        foo: 'str'
+
+
+    >>> model = FooModel()
+    >>> model.foo
+    Traceback (most recent call last):
+      ...
+    simple_model.exceptions.EmptyField: 'foo' field cannot be empty
+    """
+
+    def __getattribute__(self, name):
+        meta = object.__getattribute__(self, '_meta')
+        if name in meta.fields and not self._is_valid:
+            validate = object.__getattribute__(self, 'validate')
+            validate()
+
+        return object.__getattribute__(self, name)
+
+    def __setattr__(self, name, value):
+        meta = object.__getattribute__(self, '_meta')
+        if name in meta.fields and self._is_valid:
+            self._is_valid = False
+
+        return super().__setattr__(name, value)
+
+    def as_dict(self):
+        """
+        Returns the model as a dict
+        """
+        if not self._is_valid:
+            self.validate()
+
         from .converters import to_dict
         return to_dict(self)